Optimize Dockerfile and added Git
This started as a change to add Git so that developers will have it available if they need to install a dependency from GitHub. I then took it a step further to optimize the Dockerfile a bit for size by consolodating the `apt` commands to a single `RUN` statement and clearing the apt lists at the end. I also set it to install Rebar and not output the Hex info.