Remove Contextify by upgrading D3
Contextify was causing the Kitto container to require a number of
additional packages to build the dashboard. Contextify was no longer
required after Node.js 0.12. The version of D3 required for the
application used a very old version of JSDom which required contextify.
By upgrading D3 JSDom was upgraded and now Contextify is no longer
installed.

Rickshaw was upgraded because the version originally used required D3
3.3.

Fixes issue #8