Add webhook support
Webhooks are enabled through the `/hooks` route. Building hooks, similar
to jobs, is as easy as putting a file in the `hooks/` directory in a
Kitto project. The Hook API is very simple and similar to the
`Plug.Route` DSL building routes:

```elixir
use Kitto.Hooks.DSL

hook :github do
  {:ok, body, _} = read_body conn
  commits = GitHub.parse_commits_from_hook(body)
  broadcast! :github_commits, %{commits: commits}
end
```

The hook above generates a route at `/hooks/github` listening on all
HTTP methods. Hooks include both the `conn` object that routes include
as well as `broadcast!/2` to broadcast events to dashboards.