Beginning to rewrite the logic for defining jobs and hooks
The new implementation will replace the different registry for jobs and hooks with a single registry which manages different groups of data sources. The registry builds sub-registries, called source types, which define how data gets into a dashboard. The current source types being built in the new logic are jobs (both the Elixir and command implements) and hooks. The new implementation will also include a single DSL library to define both hooks and jobs using much of the same logic to inject helpers such as `Kitto.Notifier.Broadcast/2` into the data sources.

Redefining the implementation of DSLs and registries will help to improve test ability and coverage as well as DRY up the code.