Move hook loading to after server start
I originally built the logic to load in hooks the same way as
`Kitto.Job.Runner`, but the test kept failing, because the hooks loaded
faster than the server could start. By capturing the response of the
`GenServer.start_link` and then loading in hooks, I can assure that the
server is up.